project3/requirements.txt
streamlit
google-generativeai
groq
pypdf
python-dotenv
